+++ deepmatch-testing/src/main/kotlin/com/aouledissa/deepmatch/api/DeeplinkSpecFixtures.kt
@@ -1,26 +1,26 @@
 package com.aouledissa.deepmatch.api
 
-import kotlin.reflect.KClass
-
 object DeeplinkSpecFixtures {
 
     fun create(
+        name: String = "",
         scheme: Set<String> = setOf(),
         host: Set<String> = setOf(),
         port: Int? = null,
         pathParams: List<Param> = emptyList(),
         queryParams: Set<Param> = setOf(),
         fragment: String? = null,
-        parametersClass: KClass<out DeeplinkParams>? = null
+        paramsFactory: ((Map<String, String?>) -> DeeplinkParams?)? = null
     ): DeeplinkSpec {
         return DeeplinkSpec(
+            name = name,
             scheme = scheme,
             host = host,
             port = port,
             pathParams = pathParams,
             queryParams = queryParams,
             fragment = fragment,
-            parametersClass = parametersClass
+            paramsFactory = paramsFactory
         )
     }
 }